# DC-1091: Speed up test script by ingesting data in parallel