--- a/tools/setupResourceScripts/setup_tdr_resources.py
+++ b/tools/setupResourceScripts/setup_tdr_resources.py
@@ -40,13 +40,16 @@ def wait_for_job(clients, job_model):
             result = clients.jobs_api.retrieve_job(job_model.id)
         elif result.job_status == "failed":
             result = clients.jobs_api.retrieve_job_result(job_model.id)
-            raise f"Could not complete job with id {job_model.id}, got result {result}"
+            raise Exception(
+                f"Could not complete job with id {job_model.id}, got result {result}"
+            )
         elif result.job_status == "succeeded":
             print(f"Job succeeded {job_model.id}: {job_model.description}")
             result = clients.jobs_api.retrieve_job_result(job_model.id)
             return result
         else:
             raise "Unrecognized job state %s" % result.job_status
+
 
 def wait_for_jobs(clients, jobs):
     for job in jobs:
@@ -63,25 +66,29 @@ def convert_to_json_array(table_csv):
 
 
 def dataset_ingest_array(clients, dataset_id, dataset_to_upload):
-  jobs = []
-  for table in dataset_to_upload["tables"]:
+    jobs = []
+    for table in dataset_to_upload["tables"]:
         with open(
             os.path.join("files", dataset_to_upload["schema"], f"{table}.json")
         ) as table_csv:
-            recordsArray = convert_to_json_array(table_csv)
-            if len(recordsArray) > 0:
+            records_array = convert_to_json_array(table_csv)
+            if len(records_array) > 0:
                 ingest_request = {
                     "format": "array",
-                    "records": recordsArray,
+                    "records": records_array,
                     "table": table,
                 }
                 print(f"Ingesting data into {dataset_to_upload['name']}/{table}")
-                jobs.append(clients.datasets_api.ingest_dataset(dataset_id, ingest=ingest_request
+                jobs.append(
+                    clients.datasets_api.ingest_dataset(
+                        dataset_id, ingest=ingest_request
                     ),
                 )
             else:
-                print(f"Skipping ingest of {dataset_to_upload['name']}/{table} because it is empty")
-  wait_for_jobs(clients, jobs)
+                print(
+                    f"Skipping ingest of {dataset_to_upload['name']}/{table} because it is empty"
+                )
+    wait_for_jobs(clients, jobs)
 
 
 def find_billing_profile_by_application_deployment_name(managed_app_name):
@@ -113,9 +120,9 @@ def create_billing_profile(
         )
 
         if azure_managed_app_name:
-            billing_profile_request[
-                "applicationDeploymentName"
-            ] = azure_managed_app_name
+            billing_profile_request["applicationDeploymentName"] = (
+                azure_managed_app_name
+            )
             print(
                 f"Checking if billing profile with managed app name {azure_managed_app_name} already exists"
             )
@@ -151,25 +158,21 @@ def add_billing_profile_members(clients, profile_id):
         {"email": "DataRepoTestResourceAccess@dev.test.firecloud.org"},
     )
 
+
 def dataset_ingest_json(clients, dataset_id, dataset_to_upload):
-  jobs = []
-  for table in dataset_to_upload["tables"]:
-        with open(
-            os.path.join(
-                "files",
-                dataset_to_upload["schema"],
-                f"{table}.{dataset_to_upload['format']}",
-            )
-        ) as table_csv:
-            upload_prefix = dataset_to_upload["upload_prefix"]
-            ingest_request = {
-                "format": "json",
-                "path": f"{upload_prefix}/{table}.json",
-                "table": table,
-            }
-            print(f"Ingesting data into {dataset_to_upload['name']}/{table}")
-            jobs.append(clients.datasets_api.ingest_dataset(dataset_id, ingest=ingest_request),)
-  wait_for_jobs(clients, jobs)
+    jobs = []
+    for table in dataset_to_upload["tables"]:
+        upload_prefix = dataset_to_upload["upload_prefix"]
+        ingest_request = {
+            "format": "json",
+            "path": f"{upload_prefix}/{table}.json",
+            "table": table,
+        }
+        print(f"Ingesting data into {dataset_to_upload['name']}/{table}")
+        jobs.append(
+            clients.datasets_api.ingest_dataset(dataset_id, ingest=ingest_request),
+        )
+    wait_for_jobs(clients, jobs)
 
 
 def add_dataset_policy_members(clients, dataset_id, dataset_to_upload):
@@ -240,12 +243,12 @@ def add_snapshot_policy_members(clients, snapshot_id, snapshot_to_upload):
         clients.snapshots_api.add_snapshot_policy_member(
             snapshot_id, "discoverer", policy_member={"email": discoverer}
         )
-    for aggregateDataReader in snapshot_to_upload.get("aggregateDataReaders", []):
-        print(f"Adding {aggregateDataReader} as an aggregateDataReader")
+    for aggregate_data_reader in snapshot_to_upload.get("aggregateDataReaders", []):
+        print(f"Adding {aggregate_data_reader} as an aggregateDataReader")
         clients.snapshots_api.add_snapshot_policy_member(
             snapshot_id,
             "aggregate_data_reader",
-            policy_member={"email": aggregateDataReader},
+            policy_member={"email": aggregate_data_reader},
         )