The `setup_tdr_resources.py` script is used to initialize a dataset for local testing, or for any environment where test data is needed.

This change modifies the step where is data ingested into the dataset so that data is ingested in parallel.

before
```
python setup_tdr_resources.py --host http://localhost:8080 --datasets   0.71s user 0.23s system 0% cpu 7:53.02 total
```

after
```
python setup_tdr_resources.py --host http://localhost:8080 --datasets   0.77s user 0.28s system 0% cpu 6:02.02 total
```

so the new script is about 20% faster than the original